Update HelperMethods, Models and Forms.
Update Models by creating a new model to store a hashed password and it's corresponding salt. Rename the Password model to be StoredPassword to indicate it is a model for the passwords stored by a user and not the users password.

Update HelperMethods to add method VerifyPassword, also rename CreateHashedPassword to GenerateHashedPasswordAndSalt and update usage to populate and return the new HashedPasswordAndSalt object.

Update usages of the above in the forms