Update Tests Project, Create Test Class, and Update HelperMethods
Update the Tests project's target framework to net6.0-windows to be able to access PasswordPal.UI's HelperMethods.cs.

Create a Tests.cs to house all tests, separated by regions, add tests for all helper methods, and update database access tests to use the context and not an in-memory database.

Update HelperMethods to be more generic, making the methods more testable.