Add Help, Info, and GitHub Icons to LoginForm.
Add images for the icons and include alternate images to represent a clicked image. Update the project resources to contain the new images.

Create OnClick methods for the images, and implement functionality for the GitHub icon to link to my GitHub profile.

For the other icons (Help and Info), their functionality still needs to be implemented.