Implement NLog with logging. Update Services, Context and Tests.
Update Context to have a constructor that takes DbContextOptions as an parameter.

Update all service classes/methods that access the context to have take a context parameter that defaults to null. This is so an in-memory context can be used for testing so we can pass the specific context to the method we want to test.

Update TestDataGenerator class to user details with a Guid appended to the end rather than an random number.

Implement in-memory database for all tests.

Create TestFixture class to configure the in-memory database and seed categories.

Update all tests that use the context to pass the in-memory database to methods when testing.